# refactor: speedup URL tester