Makes getScriptsByURL faster by up to ~25%, hence also benefiting preinject. Noticeable when rebuilding all rules which happens when the cache is expired: ~48ms -> ~35ms.

It also removes the potentially memory-hungry caching of per-URL-per-regexp results: there's no need for that because V8 superbly optimizes a hot execution path (~20ms in the second run, ~5ms in the 3rd or 4th runs).